+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/sign/SKHangingSignShopObjectType.java
@@ -12,6 +12,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
 import com.nisovin.shopkeepers.api.shopobjects.sign.HangingSignShopObjectType;
+import com.nisovin.shopkeepers.compat.MC_1_20;
 import com.nisovin.shopkeepers.config.Settings;
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
@@ -38,7 +39,7 @@ public final class SKHangingSignShopObjectType
 
 	@Override
 	public boolean isEnabled() {
-		return Settings.enableHangingSignShops;
+		return MC_1_20.isAvailable() && Settings.enableHangingSignShops;
 	}
 
 	@Override